se ajusta configuracion de base de datos y se crea controller

+++ src/main/java/co/com/oospina/userhexagonalservice/application/port/output/IGetUsersByIdOutput.java
@@ -3,5 +3,5 @@ package co.com.oospina.userhexagonalservice.application.port.output;
 import co.com.oospina.userhexagonalservice.application.dto.users.UsersDto;
 
 public interface IGetUsersByIdOutput {
-    UsersDto getUsersById(Integer id);
+    UsersDto getUsersById(Long id);
 }
