+++ src/main/java/co/com/oospina/userhexagonalservice/application/services/users/IUsersById.java
@@ -3,5 +3,5 @@ package co.com.oospina.userhexagonalservice.application.services.users;
 import co.com.oospina.userhexagonalservice.application.dto.users.UsersDto;
 
 public interface IUsersById {
-    UsersDto getPostUserById(Integer id);
+    UsersDto getPostUserById(Long id);
 }
